requirements.txt
streamlit>=1.28.0
pandas>=1.5.0
plotly>=5.15.0
gspread>=5.10.0
google-auth>=2.22.0